requirements.txt
opencv-python
Pillow
transformers[agents]
diffusers 
accelerate 
datasets 
torch 
soundfile 
sentencepiece 
openai
retrying